Add Time and Modularize the classes
