+++ index.html
@@ -47,6 +47,6 @@
     <footer>
       Copyright 2022, Rights Reserved.
     </footer>
-    <script src="./index.js"></script>    
+    <script type="module" src="./index.js"></script>    
 </body>
 </html>
