+++ modules/Book.js
@@ -1,0 +1,8 @@
+class Book {
+  constructor(title, author) {
+    this.title = title;
+    this.author = author;
+  }
+}
+
+export default Book;
