+++ modules/domSelectors.js
@@ -1,0 +1,12 @@
+const domSelectors = {
+  // Nav Selectors
+  GoToBookList: document.querySelector("#book-list"),
+  GoToAddBook: document.querySelector("#add-book"),
+  GoToContact: document.querySelector("#contact"),
+  // Section Selectors
+  bookListView: document.querySelector(".sec-all-books"),
+  addBookView: document.querySelector(".sec-add-book"),
+  contactView: document.querySelector(".sec-contact-us"),
+};
+
+export default domSelectors;
