--- modules/selectors.js
@@ -1,9 +1,0 @@
-// Nav Selectors
-export const GoToBookList = document.querySelector("#book-list");
-export const GoToAddBook = document.querySelector("#add-book");
-export const GoToContact = document.querySelector("#contact");
-
-// Section Selectors
-export const bookListView = document.querySelector(".sec-all-books");
-export const addBookView = document.querySelector(".sec-add-book");
-export const contactView = document.querySelector(".sec-contact-us");
